[chore, feat, refactor] 1.3 Beta R5 C# Rewrite
# Changes
- Fix Volume typo in settings menu for default language
- Remove unused counter update function in GameManager
- Remove all references to UpdateCounter()
- Fully implement reactive Ui counter

# Fixes
- Fix the amount of coins for Abnormal Enemy

# Features
- Rewrote Settings menu Ui, Logic (Currently W.I.P: DLC downloading isn't implemented yet and Dropdown buttons don't have FancyButton class due to inheritance issues from it
- Ui, Enemy animations and Coin particles amount is now an option for settings menu
- Rewrote Upgrades, as well as their functionality in Store UI, Logic (currently W.I.P: only drooper cooldown upgrade works, and is heavily unfinished)

# Known Issues
- Drooper Income is broken due to the DrooperCooldown value being set to 0 for new savefiles
- Go to top button does nothing in changelog menu and is a stub
- Stats menu playtime shows enemies table instead
- Notifications are not showing as a visual note and aren't implemented

+++ src/UI/Menus/Changelog/ChangelogUi.cs
@@ -1,5 +1,6 @@
 using Godot;
 using TDSClicker.Core.Autoloads;
+using TDSClicker.Core.Systems;
 using TDSClicker.Utils;
 
 namespace TDSClicker.UI.Menus.Changelog;
@@ -50,21 +51,25 @@ public partial class ChangelogUi : MenuBase
 
 	private async void _switchArticle(ChangeLogData article)
 	{
-		InfoContainerTween?.Kill();
-		InfoContainerTween = CreateTween();
-		InfoContainerTween.SetTrans(Tween.TransitionType.Quad);
+		if (GameManager.Settings.UiAnimations)
+		{
+			InfoContainerTween?.Kill();
+			InfoContainerTween = CreateTween();
+			InfoContainerTween.SetTrans(Tween.TransitionType.Quad);
 		
-		InfoContainerTween.Parallel().TweenProperty(InfoContainer, "modulate:a", 0, 0.15f);
-		InfoContainerTween.Parallel().TweenProperty(InfoContainer, "position:x", 180, 0.2f);
+			InfoContainerTween.Parallel().TweenProperty(InfoContainer, "modulate:a", 0, 0.15f);
+			InfoContainerTween.Parallel().TweenProperty(InfoContainer, "position:x", 180, 0.2f);
 		
-		await ToSignal(InfoContainerTween, Tween.SignalName.Finished);
+			await ToSignal(InfoContainerTween, Tween.SignalName.Finished);
 		
-		InfoContainerTween?.Kill();
-		InfoContainerTween = CreateTween();
-		InfoContainerTween.SetTrans(Tween.TransitionType.Quad);
-		InfoContainerTween.Parallel().TweenProperty(InfoContainer, "modulate:a", 1, 0.2f);
-		InfoContainerTween.Parallel().TweenProperty(InfoContainer, "position:x", 164.0, 0.3f);
+			InfoContainerTween?.Kill();
+			InfoContainerTween = CreateTween();
+			InfoContainerTween.SetTrans(Tween.TransitionType.Quad);
+			InfoContainerTween.Parallel().TweenProperty(InfoContainer, "modulate:a", 1, 0.2f);
+			InfoContainerTween.Parallel().TweenProperty(InfoContainer, "position:x", 164.0, 0.3f);
+		}
 		InfoText.Text = article.GetLocalizedText();
-		InfoContainer.ScrollVertical = 0;
+		InfoContainer.ScrollVertical = 0;	
+
 	}
 }
